Move detailsObject to __mocks__
