+++ src/Components/PositionDetailsItem/PositionDetailsItem.test.jsx
@@ -2,7 +2,7 @@ import { shallow } from 'enzyme';
 import React from 'react';
 import toJSON from 'enzyme-to-json';
 import PositionDetailsItem from './PositionDetailsItem';
-import detailsObject from '../../Constants/TestObjects';
+import detailsObject from '../../__mocks__/detailsObject';
 
 describe('PositionDetailsItem', () => {
   let wrapper = null;
