+++ src/__mocks__/detailsObject.js
